+++ crates/cli/src/commands/install_all.rs
@@ -1,15 +1,14 @@
-use crate::helpers::{disable_progress_bars, enable_progress_bars};
+use crate::helpers::{
+    create_progress_bar, disable_progress_bars, enable_progress_bars, ToolsLoader,
+};
 use crate::{
     commands::clean::{internal_clean, CleanArgs},
     commands::install::{internal_install, InstallArgs},
-    helpers::create_progress_bar,
 };
-use futures::future::try_join_all;
 use miette::IntoDiagnostic;
-use proto_core::{load_tool_from_locator, ProtoEnvironment, ToolsConfig, UserConfig};
 use starbase::system;
 use starbase_styles::color;
-use std::env;
+use std::{env, process};
 use tracing::{debug, info};
 
 #[system]
@@ -18,69 +17,73 @@ pub async fn install_all() {
 
     debug!("Loading tools and plugins from .prototools");
 
-    let user_config = UserConfig::load()?;
-    let mut config = ToolsConfig::load_upwards()?;
-    config.inherit_builtin_plugins();
+    let loader = ToolsLoader::new()?;
+    let tools = loader.load_tools().await?;
 
     debug!("Detecting tool versions to install");
 
-    let proto = ProtoEnvironment::new()?;
+    let mut versions = loader.tools_config.tools.clone();
 
-    for (name, locator) in config.plugins {
-        if config.tools.contains_key(&name) {
+    for tool in &tools {
+        if versions.contains_key(&tool.id) {
             continue;
         }
-
-        let tool = load_tool_from_locator(&name, &proto, &locator, &user_config).await?;
 
         if let Some(candidate) = tool.detect_version_from(&working_dir).await? {
             debug!("Detected version {} for {}", candidate, tool.get_name());
 
-            config.tools.insert(name, candidate);
+            versions.insert(tool.id.clone(), candidate);
         }
     }
 
-    if config.tools.is_empty() {
-        info!("Nothing to install!");
-    } else {
-        let pb = create_progress_bar(format!(
-            "Installing {} tools: {}",
-            config.tools.len(),
-            config
-                .tools
-                .keys()
-                .map(color::id)
-                .collect::<Vec<_>>()
-                .join(", ")
-        ));
+    if versions.is_empty() {
+        eprintln!("Nothing to install!");
+        process::exit(1);
+    }
 
-        disable_progress_bars();
+    let pb = create_progress_bar(format!(
+        "Installing {} tools: {}",
+        versions.len(),
+        versions
+            .keys()
+            .map(color::id)
+            .collect::<Vec<_>>()
+            .join(", ")
+    ));
 
-        let mut futures = vec![];
+    disable_progress_bars();
 
-        for (id, version) in config.tools {
+    let mut futures = vec![];
+
+    for tool in tools {
+        if let Some(version) = versions.remove(&tool.id) {
             futures.push(tokio::spawn(async {
-                internal_install(InstallArgs {
-                    canary: false,
-                    id,
-                    pin: false,
-                    passthrough: vec![],
-                    spec: Some(version),
-                })
+                internal_install(
+                    InstallArgs {
+                        canary: false,
+                        id: tool.id.clone(),
+                        pin: false,
+                        passthrough: vec![],
+                        spec: Some(version),
+                    },
+                    Some(tool),
+                )
                 .await
             }));
         }
-
-        try_join_all(futures).await.into_diagnostic()?;
-
-        enable_progress_bars();
-
-        pb.finish_and_clear();
-
-        info!("Successfully installed tools");
     }
 
-    if user_config.auto_clean {
+    for future in futures {
+        future.await.into_diagnostic()??;
+    }
+
+    enable_progress_bars();
+
+    pb.finish_and_clear();
+
+    info!("Successfully installed tools");
+
+    if loader.user_config.auto_clean {
         debug!("Auto-clean enabled, starting clean");
 
         internal_clean(&CleanArgs {