+++ crates/core/tests/version_resolver_test.rs
@@ -120,7 +120,7 @@ mod version_resolver {
     }
 
     #[test]
-    #[should_panic(expected = "Failed to resolve a semantic version for unknown.")]
+    #[should_panic(expected = "Failed to resolve unknown to a valid supported version.")]
     fn errors_unknown_alias() {
         let versions = create_versions();
         let aliases = create_aliases();
@@ -135,7 +135,7 @@ mod version_resolver {
     }
 
     #[test]
-    #[should_panic(expected = "Failed to resolve a semantic version for missing.")]
+    #[should_panic(expected = "Failed to resolve missing to a valid supported version.")]
     fn errors_missing_alias_from_alias() {
         let versions = create_versions();
         let aliases = create_aliases();
@@ -150,7 +150,7 @@ mod version_resolver {
     }
 
     #[test]
-    #[should_panic(expected = "Failed to resolve a semantic version for 20.0.0.")]
+    #[should_panic(expected = "Failed to resolve 20.0.0 to a valid supported version.")]
     fn errors_missing_version_from_alias() {
         let versions = create_versions();
         let aliases = create_aliases();
@@ -307,7 +307,7 @@ mod version_resolver {
     }
 
     #[test]
-    #[should_panic(expected = "Failed to resolve a semantic version for 20.0.0.")]
+    #[should_panic(expected = "Failed to resolve 20.0.0 to a valid supported version.")]
     fn errors_unknown_version() {
         let versions = create_versions();
         let aliases = create_aliases();
@@ -396,7 +396,7 @@ mod version_resolver {
     }
 
     #[test]
-    #[should_panic(expected = "Failed to resolve a semantic version for ^20.")]
+    #[should_panic(expected = "Failed to resolve ^20 to a valid supported version.")]
     fn errors_no_req_match() {
         let versions = create_versions();
         let aliases = create_aliases();
@@ -428,7 +428,7 @@ mod version_resolver {
     }
 
     #[test]
-    #[should_panic(expected = "Failed to resolve a semantic version for ^9 || ^5 || ^3.")]
+    #[should_panic(expected = "Failed to resolve ^9 || ^5 || ^3 to a valid supported version.")]
     fn errors_no_req_any_match() {
         let versions = create_versions();
         let aliases = create_aliases();